split to request_parser and response_writer for single responsive
